// src/adapter/fm-data/interfaces/Registries/Registry.ts
/* --------
 * Registry
 * -------- */
export interface Registry {
  IdAccount: string;

  TxtAccount: string;

  TxtCognome: string;

  TxtNome: string;

  MailAziendale: string;

  CodiceAccount: string;

  CodiceFiscale: string;

  DescrizioneGenerica: string;

  DescrizioneLegale: string;

  FlagCompagnia: number;

}